+++ .gitignore
@@ -2,3 +2,4 @@ __pycache__/
 ytd/__pycache__/operators.cpython-310.pyc
 __pycache__/vicho_misc.cpython-310.pyc
 ytd/__pycache__/ytd_helper.cpython-310.pyc
+__pycache__/__init__.cpython-310.pyc
